base: clean up redundant casts
Some `as usize` casts are necessary on non-glibc C libraries, since the
type of `cmsg_len` and `msg_controllen` may not be usize in the libc
bindings, so these warnings are suppressed instead of removing the
casts.

BUG=b:276487055
TEST=tools/clippy # with Rust 1.68

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4391102
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>

+++ base/src/sys/unix/process.rs
@@ -36,7 +36,7 @@ impl Child {
         if let Some(exit_code) = status.code() {
             Ok(exit_code as u8)
         } else if let Some(signal) = status.signal() {
-            let exit_code = if signal as i32 >= 128 {
+            let exit_code = if signal >= 128 {
                 warn!("wait for child: unexpected signal({:?})", signal);
                 255
             } else {
